Open an edit view to a form using direct URL
User can now directly open an edit view of a form using direct URL. The
URL pattern is: /lomake-editori/#/editor/forms/:id where :id is the ID
of the form to be edited.

This commit also removes the :fetch-initial-data event dispatching from
lomake-editori.core/init because that event prevents anything else to happen
whenever a request comes in and events are dispatched at lomake-editori.routes.