Take ca-initiated logout into account
CAS-initiated logout happens when logout is done outside of our appication.
In this case CAS POSTS to a url in our system. We have no user session when this happens
because browser didn't initiate this call, so we have to work with the passed-in ticket instead.
This ticket will be invalidated in store, and logged-in check must check that the ticket is still valid.
This version uses in-memory store, but next there will be a DB version.