Fix IE bug
IE launched multiple events from the events vector (MSAnimationEnd, animationend). Therefore we removed two elements with the index we calculated. In practice this caused multiple elements to be removed when multiple events occurred.

animationend seems to be widely supported, so let's err on the side of safety and not remove excess elements.