Fix implementation based on new knowledge about real data
I had to look into real data in order to figure out that
previous implementation was immune to the type of key in
liitepyynnot-for-hakukohteet. The key is keyword and the implementation
was transforming it again into keyword and clojure is not complaining about that!
So (keyword :abc) is happily again :abc